backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
beautifulsoup4==4.12.3
typing-extensions==4.12.2

